Fix regression in 0aa7ec933bfa77193aac7a.
When I ran go test ./... locally, I failed to notice the one liner at the top
signaling this error.

TBR=tandrii@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1157293003

+++ client/isolate/isolate_test.go
@@ -179,7 +179,7 @@ func TestArchive(t *testing.T) {
 
 	ut.AssertEqual(t, nil, server.Error())
 	digest, err := isolated.HashFile(filepath.Join(tmpDir, "baz.isolated"))
-	ut.AssertEqual(t, isolated.DigestItem{isolatedHash, false, len(isolatedEncoded)}, digest)
+	ut.AssertEqual(t, isolated.DigestItem{isolatedHash, false, int64(len(isolatedEncoded))}, digest)
 	ut.AssertEqual(t, nil, err)
 }
